implement the `start` lang item
with this one can use the normal form of `main` in executables, i.e. `fn
main() { .. }`

this commit also re-exports some macros from the `core` crate

closes #14
