+++ examples/echo.rs
@@ -1,17 +1,9 @@
-#![no_std]
-#![no_main]
-
-#[macro_use]
-extern crate std;
-
 use std::io;
 use std::io::{Read, Write};
 
-#[no_mangle]
-pub fn main() -> i32 {
+fn main() {
     let mut buffer = [0; 256];
     let mut stdin = io::stdin();
     let n = stdin.read(&mut buffer).unwrap();
     io::stdout().write_all(&buffer[..n]).unwrap();
-    0
 }
