+++ examples/format.rs
@@ -1,11 +1,3 @@
-#![no_std]
-#![no_main]
-
-#[macro_use]
-extern crate std;
-
-#[no_mangle]
-pub fn main() -> i32 {
+fn main() {
     println!("{} + {} = {}", 1, 1, 1 + 1);
-    0
 }
