+++ examples/instant.rs
@@ -1,16 +1,8 @@
-#![no_std]
-#![no_main]
-
-#[macro_use]
-extern crate std;
-
 use std::time::Instant;
 
-#[no_mangle]
-pub fn main() -> i32 {
+fn main() {
     let now = Instant::now();
     let elapsed = now.elapsed();
 
     println!("{:?}", elapsed);
-    0
 }
