+++ examples/panic.rs
@@ -1,9 +1,3 @@
-#![no_std]
-#![no_main]
-
-extern crate std;
-
-#[no_mangle]
-pub fn main() -> i32 {
+fn main() {
     panic!()
 }
