+++ examples/system-time.rs
@@ -1,15 +1,7 @@
-#![no_std]
-#![no_main]
-
-#[macro_use]
-extern crate std;
-
 use std::time::SystemTime;
 
-#[no_mangle]
-pub fn main() -> i32 {
+fn main() {
     let now = SystemTime::now();
 
     println!("{:?}", now);
-    0
 }
