+++ examples/zero.rs
@@ -1,9 +1,1 @@
-#![no_std]
-#![no_main]
-
-extern crate std;
-
-#[no_mangle]
-pub fn main() -> i32 {
-    0
-}
+fn main() {}
